Configure CircleCI builds to deploy code-club-world branch to GH Pages
This is specific to the RPF CodeClubWorld app and should not be merged
into the main "develop" branch or pushed to the upstream repo.

I don't think we need the deploy-npm or push-translations workflows.

The `npm run deploy -- -e $CIRCLE_BRANCH` step in the deploy-gh-pages
job appears to make use of the gh-pages `-e` option [1] to publish the
changes to the gh-pages branch within a sub-directory named after the
current branch. I've retained this functionality given that it may well
be useful if we want to try out changes without affecting production.

[1]: https://github.com/rschamp/gh-pages/commit/e0dd4eaa54276a8dac2be16fd4c75a233bdd582c#diff-558572b6f7bd7068bbbd3db502651a5a313bfc3901d72eb2995833592e4434edR30