feat: Allow MenuBar to be hidden
The MenuBar is quite specific to the canonical version of the online
Scratch editor [1]. We want to make use of the scratch-gui package
without displaying the MenuBar.

Since the height of the menu bar is hard-coded as a CSS variable, I've
had to override the height calculation for the body wrapper when the
menu bar is not visible. Ideally the height of the menu bar wouldn't be
hard-coded and this wouldn't be necessary.

[1]: https://scratch.mit.edu/projects/editor