Add top-level HTML page to render GUI component for Code Club World
These changes result in a new code-club-world.html page being made
available alongside other similar pages. The idea is still that this
page is embedded in an iframe within a page served by the Rails app.

The new page is closely based on the existing index.html page which we
have been using. However, a number of properties are set specifically
for the Code Club World application (see below).

If we could publish our fork of the scratch-gui package (or ideally get
the relevant changes incorporated into the canonical package), there's
no reason why this page needs to live in this repo. The only reason I've
added it here for now is for convenience since we already have a CI
pipeline set up to deploy to GitHub Pages.

* The new menuBarHidden property (added in an earlier commit) is set to
true to hide the MenuBar component.

* The projectHost & assetHost properties are set to point to endpoints
provided by the Rails app which is the app that serves the parent HTML
page.

* If no projectId parameter is present in the URL of the editor page, then
the GUI component projectId property is set to 0. This causes the GUI
component to build a new default Scratch project.

* The fact that the canCreateNew property is set to true means that a
newly built project will immediately be saved using the projectHost
endpoint via a POST request. Setting this canCreateNew property
effectively simulates the logged-in state in the canonical version of
the editor hosted by MIT.

* If a projectId parameter is present in the URL of the editor page,
then the GUI component projectId property is set to this value. This
causes the GUI component to load the project from the projectHost
endpoint via a GET request.

* The fact that the canSave property is set to true and the
autoSaveIntervalSecs property is set to 10 means that if the user makes
any changes after the project has been created or loaded then 10 seconds
later the project will be saved via a PUT request.

Some differences from index.html (and src/playground/render-gui.jsx):

* Detection of unsupported browsers not included (see
BrowserModalComponent). We can always revisit this later, but I wanted
to keep the code as simple as possible for now.  I have included the
same polyfills, because that seemed more useful and not much code.

* No simulation of Scratch Desktop.

* No Backpack configured.

* No warning before navigating away. We may well want to address this,
but we'll probably need to tackle it in the parent page served by the
Rails app.

* The HashParserHOC component has not been included, because it was
messing with the ability to get the GUI component to create a default
project when projectId was set to 0. HashParserHOC seems to provide some
(legacy?) functionality which allows the project ID to be specified in
the URL hash, but we don't need that so we can safely leave it out.