Include origin, path & query string in Referer for requests
The projectHost & assetHost endpoints provided by the parent Rails app
need to know which Code Club World project a given Scratch program is
being saved against. Rather than patching the scratch-gui app to provide
this information in the request, we can relax the Referrer Policy [1]
for the page slightly to allow these Referer values to be sent even
though the endpoints are on a different sub-domain and thus have a
different origin. The default policy is
"strict-origin-when-cross-origin" which would not allow this.

This obviously isn't ideal from a security point-of-view, but we do own
both domains, so it doesn't seem so bad. Also we have some ideas about
how we could host this page within the Rails app itself and thus have it
on the same domain. That would mean we could dispense with this policy
change.

Note that since this change is in the EJS template it will affect all
the top-level HTML pages.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Referrer-Policy