Send cookies in GET request to projectHost endpoint
The projectHost endpoint provided by the parent Rails app needs to have
access to both the session cookie and the anonymous_user_id cookie so
that it knows which Code Club World user the Scratch program needs to be
saved against.

Currently those endpoints are on a different subdomain and thus a
different origin. This means the underlying call the the Fetch API needs
to have its credentials option set to "include" [1].

The create (POST) and update (PUT) endpoints already have this
configured via getProjectCreateConfig & getProjectUpdateConfig functions
because they set the withCredentials option to true [2]. While the
scratch-storage package doesn't support the same withCredentials option
for the getProjectGetConfig function [3,4], it does allow arbitrary
options to be passed to the underlying call to the Fetch API and so we
can achieve the desired effect by returning an object with a url &
credentials properties instead of just a String.

Ideally we'd open a PR against the scratch-storage package to provide
the same withCredentials option for getProjectGetConfig so we could use
the same style as getProjectCreateConfig & getProjectUpdateConfig.

Slightly more speculatively we could also open a PR against the
scratch-gui package to allow the GUI component to be configured with a
custom Storage class.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/Fetch_API/Using_Fetch#sending_a_request_with_credentials_included
[2]: https://github.com/RaspberryPiFoundation/scratch-gui/blob/735993143ab848c6749a83f116bd5b218077c0fa/src/lib/storage.js#L41-L52
[3]: https://github.com/LLK/scratch-storage/blob/fbce0ee3707807015cd463136b28f7d91b65dd2e/src/FetchTool.js#L43
[4]: https://github.com/LLK/scratch-storage/blob/fbce0ee3707807015cd463136b28f7d91b65dd2e/src/FetchTool.js#L22