Export saveProjectToServer & add support for extra HTTP headers
I want to be able to send an X-CSRF-Token HTTP header when sending a
POST or PUT request to create or update the project.

* Exporting saveProjectToServer means that I can set the
onUpdateProjectData property to a function which wraps
saveProjectToServer.

* Adding the optional headers parameter to the saveProjectToServer
function means that the wrapping function can pass in extra HTTP headers
to be sent in any request.