Publish scoped package on successful builds
This reinstates the deploy-npm CircleCI job and tweaks it slightly:

* Scope the package name by organisation. This ties the package to the
RPF organisation rather than just to my user and means that, for
example, any member of the organisation can publish the package.

* Only publish package on successful builds of the code-club-world
branch.

* Use organisation-specific release version and git tag to avoid
confusion with versions in the canonical package and tags in upstream
repo.

* Always publish using "latest" npm tag; not "hotfix" or "stable".

* Add `--access public` option to `npm publish` command, because scoped
packages only allow "restricted" access by default.

* I've added the access token from my npmjs.org account as the NPM_TOKEN
env var in the CircleCI project settings.

* Add ssh key fingerprint to allow `git push`.