Fix proptype validation warning in SpriteList
We've seen a PropType warning for this prop in development. It seems
that when the SpriteList component is used to manage the costumes in
the left hand side of the Costumes tab, numeric identifiers are used
for `sprite`. When the SpriteList is used in the sprite selection area
under the stage, strings are used. There's probably a way of fixing
this at source, but for now being a bit more permissive with the
validation removes the warning. As far as I can tell from reading the
source, we use this id for comparison with the `===` operator, so I
think numbers or strings are functionally equivalent.