refactor: remove unused STATIC_PATH from webpack config
STATIC_PATH was introduced back in 2019 [1] as a way to allow consumers of
scratch-gui (like ourselves) to customize the publicPath (where assets are
served from when the JS is run in the browser).

However it looks like this feature was removed/commented out [2] when they
migrated to using the shared webpack config (from the npm package
`scratch-webpack-configuration`).

We would like to re-introduce this feature in order to fix an issue we have
noticed with image assets failing to load as the browser is attempting to load
thse assets from URLs relative to the webpack entry point (rather than from the
root or `/scratch-gui`).

Therefore I am removing this commented out code to make the existing config less
confusing.

[1] https://github.com/scratchfoundation/scratch-gui/commit/e2a2dcf0f
[2] https://github.com/scratchfoundation/scratch-gui/commit/f1438f73e

Co-Authored-By: Chris Lowis <chris.lowis@gofreerange.com>