refactor: group asset loading webpack config together
The existing webpack config customizes the filenames of any assets we import
which have not been inlined using the `assetModuleFilename`.

This config option is defined at the global level, however it only applies to
assets that are processed by the Asset module definition.

Rather than configuring this globally, I have opted for defining it within the
Asset module definition (nested under the `generator` option) as that makes it
clearer that this option is only relevant for those assets thus making it easier
for future developers to understand and make changes in the future, especially
for those who are less familiar with the intricacies of webpack config.

[1] https://webpack.js.org/guides/asset-modules/

Co-Authored-By: Chris Lowis <chris.lowis@gofreerange.com>